+++ src/OpenTelemetry/Metrics/IMetricsListener/OpenTelemetryMetricsListener.cs
@@ -1,7 +1,6 @@
 // Copyright The OpenTelemetry Authors
 // SPDX-License-Identifier: Apache-2.0
 
-using System.Diagnostics;
 using System.Diagnostics.Metrics;
 using Microsoft.Extensions.Diagnostics.Metrics;
 
@@ -14,14 +13,9 @@ internal sealed class OpenTelemetryMetricsListener : IMetricsListener, IDisposab
 
     public OpenTelemetryMetricsListener(MeterProvider meterProvider)
     {
-        this.meterProviderSdk = meterProvider as MeterProviderSdk;
-
-        Debug.Assert(this.meterProviderSdk != null, "meterProvider was not MeterProviderSdk");
-
-#pragma warning disable CA1508 // Avoid dead conditional code
-        if (this.meterProviderSdk != null)
-#pragma warning restore CA1508 // Avoid dead conditional code
+        if (meterProvider is MeterProviderSdk meterProviderSdk)
         {
+            this.meterProviderSdk = meterProviderSdk;
             this.meterProviderSdk.OnCollectObservableInstruments += this.OnCollectObservableInstruments;
         }
     }
@@ -30,53 +24,46 @@ internal sealed class OpenTelemetryMetricsListener : IMetricsListener, IDisposab
 
     public void Dispose()
     {
-        if (this.meterProviderSdk != null)
+        if (this.meterProviderSdk is { })
         {
             this.meterProviderSdk.OnCollectObservableInstruments -= this.OnCollectObservableInstruments;
         }
     }
 
-    public MeasurementHandlers GetMeasurementHandlers()
+    public MeasurementHandlers GetMeasurementHandlers() => new()
     {
-        return new MeasurementHandlers()
-        {
-            ByteHandler = (instrument, value, tags, state)
-                => MeterProviderSdk.MeasurementRecordedLong(instrument, value, tags, state),
-            ShortHandler = (instrument, value, tags, state)
-                => MeterProviderSdk.MeasurementRecordedLong(instrument, value, tags, state),
-            IntHandler = (instrument, value, tags, state)
-                => MeterProviderSdk.MeasurementRecordedLong(instrument, value, tags, state),
-            LongHandler = MeterProviderSdk.MeasurementRecordedLong,
-            FloatHandler = (instrument, value, tags, state)
-                => MeterProviderSdk.MeasurementRecordedDouble(instrument, value, tags, state),
-            DoubleHandler = MeterProviderSdk.MeasurementRecordedDouble,
-        };
-    }
+        ByteHandler = (instrument, value, tags, state)
+            => MeterProviderSdk.MeasurementRecordedLong(instrument, value, tags, state),
+        ShortHandler = (instrument, value, tags, state)
+            => MeterProviderSdk.MeasurementRecordedLong(instrument, value, tags, state),
+        IntHandler = (instrument, value, tags, state)
+            => MeterProviderSdk.MeasurementRecordedLong(instrument, value, tags, state),
+        LongHandler = MeterProviderSdk.MeasurementRecordedLong,
+        FloatHandler = (instrument, value, tags, state)
+            => MeterProviderSdk.MeasurementRecordedDouble(instrument, value, tags, state),
+        DoubleHandler = MeterProviderSdk.MeasurementRecordedDouble,
+    };
 
     public bool InstrumentPublished(Instrument instrument, out object? userState)
     {
-        if (this.meterProviderSdk == null)
+        userState = null;
+        bool result = false;
+
+        if (this.meterProviderSdk is { })
         {
-            userState = null;
-            return false;
+            userState = this.meterProviderSdk.InstrumentPublished(instrument, listeningIsManagedExternally: true);
+            result = userState is not null;
         }
 
-        userState = this.meterProviderSdk.InstrumentPublished(instrument, listeningIsManagedExternally: true);
-        return userState != null;
+        return result;
     }
 
     public void MeasurementsCompleted(Instrument instrument, object? userState)
-    {
-        MeterProviderSdk.MeasurementsCompleted(instrument, userState);
-    }
+        => MeterProviderSdk.MeasurementsCompleted(instrument, userState);
 
     public void Initialize(IObservableInstrumentsSource source)
-    {
-        this.observableInstrumentsSource = source;
-    }
+        => this.observableInstrumentsSource = source;
 
     private void OnCollectObservableInstruments()
-    {
-        this.observableInstrumentsSource?.RecordObservableInstruments();
-    }
+        => this.observableInstrumentsSource?.RecordObservableInstruments();
 }