+++ src/OpenTelemetry.Extensions.Hosting/CHANGELOG.md
@@ -6,6 +6,9 @@ Notes](../../RELEASENOTES.md).
 
 ## Unreleased
 
+* Fix NullReferenceException when `OTEL_SDK_DISABLED=true`.
+  ([#6869](https://github.com/open-telemetry/opentelemetry-dotnet/pull/6869))
+
 ## 1.15.0
 
 Released 2026-Jan-21